fix contract argument and add e2e test
